A project for OOP topic repetition
This project follows the decrease or increase in the number of animals such as wolves, sheep, cows, hunters, etc. in a 100*100 area (we can change this value by changing the area section in the code) according to their distance from each other. If there is a hunter-hunter relationship, the prey is eaten, if there are different animal breeds belonging to the same species, they mate and give birth to new animals. While doing this project, I wrote and designed the classes, constructors and other structures according to OOP principles. In this way, if desired, we can transform the code into different situations with small changes, add new features, new rules, change the number of animals and even new animals. I often used structures such as java Random arraylist list hashmap map in the project. I also used these structures to optimize the code. Thanks to Random, situations such as the gender of the movement directions at the time of birth are determined completely randomly, by chance. I also optimized the code thanks to structures such as List Map that I used. I made it work better when there were more animals and more rounds in the loop.
In short, I think it was a nice project that allowed me to do a nice OOP topic repetetion.